Minor refactor
Converted all methods to ES6 arrow syntax for consistent styling all the way
through the app. Since we aren't using `this` anywhere, there is no problem
with the bindings.

Add “error handling” to onResponse method. Just printed and returned the error
passed to it.

Modify lines array by adding a filter for trimmed strings being greater than 0
length. This removes the need to process these lines.

Turned `lineType`s into “enum” with an object. This makes it more obvious this
is an “enum” type and it sections those variables.

Turned the `forEach` loop in the `parseLines` function into a for loop. This
was done because the original `forEach` loop needed special checks on each
line to determine its validity as good information. With the previous change
of filtering the lines array by line length in addition to this change, we
can be confident that every line that is read by this loop is usable
information. The reason for starting the iteration at 3 is because the first
two lines are table headers (index 0 and 1) and the third line is a splitter
(index 2).

With the above change, the line still needed special characters replaced. This
is done here with regex replacement of `&amp;` and `&apos;` to `&` and `’` to
make sure all lines are the same length - these characters were being escaped
into their HTML entities which was throwing off the parsing.

If statements for determining `lineType` was turned into an if / else branch.
This is because a line cannot be more than one lineType, and if the lineType
has been found, another will not take its place. By using an if / else branch,
the script does not have to check each individual if statement anymore. Once it
finds the answer, it can continue to the next steps.

The last if / else branch for determining what to do with `lineType` enum has
been turned into a switch statement.

The resulting object from `parseIntervalLine` has had some attributes
implicitly defined. Example: `hasTime:  hasTime,` was turned into `hasTime,`.

Prettier was ran. I'm not sure this should become a habit yet because it seems
to still be pretty buggy and doesn't like some styling that is in this file.