Separated into composable scrapers, added department name scraping
This commit separates scrapers into composable units that return promises and can be chained. They feed data into one another and depend on the data of the previous scraper promise in order to make alterations to it. They pass that object along until each scraper is done making alterations to it

This also adds a composable scrape section that scrapes department names and semester module names from the selection dropdowns to ensure that we are parsing all the correct departments. Semester modules names are not currently used but may be used in the future to store data for multiple semesters.