Aggregate teachers, fix teacher name bug
There was an additional bug with parsing teachers names. The last letter (first initial) was cut off due to incorrect slice bounds. This is now fixed.

Also the teachers are now aggregated and sorted for uniques. I will use this in a strategy to create unique database entries for teachers but then also link them to their courses.